Add a bundle distribution.
Using Quicklisp bundle facility it is possible to prepare a
self-contained archive of all the code needed to build pgloader.

Doing that should allow users to easily build pgloader when they are
being a restrictive proxy, and packagers to work from a source tarball
that has a very limited build dependencies.