Add a CCL dockerfile.
For some reasons, with the default DYNSIZE and even when using the 64
bits Clozure-CL variant, I get a series of error messages like the one
below, so that I had to restrain to using 256 MB only:

  Fatal error in "buildapp" : Fault during read of memory address #x7F8C37522668
  Fatal error in "buildapp" : Fault during
  Fatal error in "buildapp" : Stack overflow on temp stack.
  Fatal error in "buildapp" : Fault during read of memory address #x7F8C37522668

It's worth trying something else as the limitation might be tied to my
local virtual build environment.

See #327 where the SBCL Garbage Collector is introducing problems which
might not appear at all when compiling with Clozure-CL instead.